Renames and fixed pipeline.
