+++ .github/workflows/release.yml
@@ -25,7 +25,7 @@ jobs:
       - name: Setup .NET
         uses: actions/setup-dotnet@v1
         with:
-          dotnet-version: 9.0.x
+          dotnet-version: 10.0.x
 
       # Restore projects
       - name: Restore
