+++ tests/Sourcey.Tests/Aggregates/Aggregate/Guards/WhenNullArguments.cs
