+++ tests/Sourcey.Tests/Aggregates/Aggregate/History/WhenReplayingAndClearingUncommitted.cs
