+++ tests/Sourcey.Tests/Events/Streams/WhenUsingEventStream.cs
