+++ tests/Sourcey.Tests/Keys/WhenUsingValueObjects.cs
